+++ packages/app/public/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+### Fixed
+- TFT has no effect on the highest league rank in the overlay anymore.
+
 ## [4.3.0] - 2019-09-01
 ### Added
 - New trees for Hecarim, JarvanIV, Poppy, Sylas, Vladimir